requirements.txt
attrs
configparser
argparse
requests